requirements.txt
pydantic
python-dotenv
rich
jinja2
requests
playwright
beautifulsoup4
pyyaml